requirement.txt
lpips==0.1.4
matplotlib==3.5.2
numpy==1.21.0
pandas==2.0.3
Pillow==9.1.1
pytorch-wavelets==1.3.0
PyWavelets==1.4.1
scipy==1.8.1
seaborn==0.12.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
